+++ jannovar-vardbs/src/test/java/de/charite/compbio/jannovar/vardbs/clinvar/ClinVarVariantContextToRecordConverterTest.java
@@ -31,25 +31,27 @@ public class ClinVarVariantContextToRecordConverterTest {
 
 	@Test
 	public void test() {
-		ClinVarVariantContextToRecordConverter converter = new ClinVarVariantContextToRecordConverter();
+		ClinVarVariantContextToRecordConverter converter =
+				new ClinVarVariantContextToRecordConverter();
 		CloseableIterator<VariantContext> it = vcfReader.iterator();
 		VariantContext vc = null;
 		while (vc == null || !vc.getID().equals("rs387907305"))
 			vc = it.next();
 
 		ClinVarRecord record = converter.convert(vc);
-		Assert.assertEquals("ClinVarRecord [chrom=1, pos=2160305, id=rs387907305, ref=G, alt=[A, T], "
-				+ "annotations={1=[ClinVarAnnotation [alleleMapping=1, hgvsVariant=NC_000001.10:g.2160306G>A, "
-				+ "sourceInfos=[ClinVarAnnotationSourceInfo [dbName=OMIM_Allelic_Variant, dbId=164780.0003], "
-				+ "ClinVarAnnotationSourceInfo [dbName=UniProtKB_(protein), dbId=P12755#VAR_071175]], origin=[BIPARENTAL], "
-				+ "diseaseInfos=[ClinVarDiseaseInfo [significance=PATHOGENIC, diseaseDB=MedGen:OMIM:SNOMED_CT, "
-				+ "diseaseDBID=C1321551:182212:83092002, diseaseDBName=Shprintzen-Goldberg_syndrome, revisionStatus=SINGLE, "
-				+ "ClinicalAccession=RCV000030818.28]]]], 2=[ClinVarAnnotation [alleleMapping=2, hgvsVariant=NC_000001.10:g.2160306G>T, "
-				+ "sourceInfos=[ClinVarAnnotationSourceInfo [dbName=OMIM_Allelic_Variant, dbId=164780.0007], "
-				+ "ClinVarAnnotationSourceInfo [dbName=UniProtKB_(protein), dbId=P12755#VAR_071177]], origin=[UNKNOWN], "
-				+ "diseaseInfos=[ClinVarDiseaseInfo [significance=PATHOGENIC, diseaseDB=MedGen:OMIM:SNOMED_CT, "
-				+ "diseaseDBID=C1321551:182212:83092002, diseaseDBName=Shprintzen-Goldberg_syndrome, revisionStatus=NO_CRITERIA, "
-				+ "ClinicalAccession=RCV000033005.27]]]]}]", record.toString());
+		Assert.assertEquals(
+				"ClinVarRecord [chrom=1, pos=2160305, id=rs387907305, ref=G, alt=[A, T], annotations={1=[ClinVarAnnotation "
+						+ "[alleleMapping=1, hgvsVariant=NC_000001.10:g.2160306G>A, sourceInfos=[ClinVarAnnotationSourceInfo "
+						+ "[dbName=OMIM_Allelic_Variant, dbId=164780.0003], ClinVarAnnotationSourceInfo [dbName=UniProtKB_(protein), "
+						+ "dbId=P12755#VAR_071175]], origin=[DE_NOVO], diseaseInfos=[ClinVarDiseaseInfo [significance=PATHOGENIC, "
+						+ "diseaseDB=MedGen:OMIM:SNOMED_CT, diseaseDBID=C1321551:182212:83092002, diseaseDBName=Shprintzen-Goldberg_syndrome, "
+						+ "revisionStatus=SINGLE, ClinicalAccession=RCV000030818.28]]]], 2=[ClinVarAnnotation [alleleMapping=2, "
+						+ "hgvsVariant=NC_000001.10:g.2160306G>T, sourceInfos=[ClinVarAnnotationSourceInfo [dbName=OMIM_Allelic_Variant, "
+						+ "dbId=164780.0007], ClinVarAnnotationSourceInfo [dbName=UniProtKB_(protein), dbId=P12755#VAR_071177]], "
+						+ "origin=[UNKNOWN], diseaseInfos=[ClinVarDiseaseInfo [significance=PATHOGENIC, diseaseDB=MedGen:OMIM:SNOMED_CT, "
+						+ "diseaseDBID=C1321551:182212:83092002, diseaseDBName=Shprintzen-Goldberg_syndrome, revisionStatus=NO_CRITERIA, "
+						+ "ClinicalAccession=RCV000033005.27]]]]}]",
+				record.toString());
 	}
 
 }