+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## develop (unreleased)
 
+### overall
+
+* Changing HTSJDK version to 2.11.0
+
+### jannovar-vardbs
+
+* Fixed problem with interpretation of Clinvar annotation origin.
+
 ## v0.23
 
 ### overall